requirements.txt
# ================================
# Web Scraping (核心抓取能力)
# ================================

scrapy>=2.11.0
# 成熟的爬虫框架
# - 内置请求调度、重试、并发控制
# - response.xpath() 底层基于 lxml
# - 适合规则化、长期运行的抓取任务

playwright>=1.40.0
# 浏览器自动化（Chromium / WebKit / Firefox）
# - 用于 JS 渲染页面
# - 解决动态加载、反爬 JS 的站点
# - 与 XPath 解析配合使用

requests>=2.31.0
# 同步 HTTP 客户端
# - 轻量、稳定
# - 适合简单页面或一次性请求

httpx>=0.25.0
# 现代 HTTP 客户端（支持 async / sync）
# - 为后续异步抓取或高并发预留
# - 与 asyncio / trio 生态兼容


# ================================
# HTML / XML Parsing (XPath 核心)
# ================================

lxml>=4.9.3
# XPath / XSLT 引擎
# - XPath 1.0 标准实现
# - 性能高、稳定
# - 银行 / 政府类网站解析首选

beautifulsoup4>=4.12.0
# HTML 解析（bs4）
# - 简单灵活的 DOM 操作
# - 用于 CSS 选择器提取内容


# ================================
# RSS Feed 支持
# ================================

feedparser>=6.0.10
# RSS / Atom 解析
# - 适合官网、媒体发布源
# - 可绕过部分反爬限制
# - 与网页抓取并行使用


# ================================
# Task Scheduling (定时任务)
# ================================

apscheduler>=3.10.4
# Python 定时任务框架
# - cron 风格调度
# - 适合本地或服务器长期运行
# - 可替代系统 cron（可选）


# ================================
# Configuration & Data
# ================================

pyyaml>=6.0.1
# YAML 配置解析
# - 存放站点规则、XPath 模板
# - 比 JSON 更适合人工维护

python-dotenv>=1.0.0
# 环境变量管理
# - 管理 token、推送密钥
# - 避免敏感信息进代码库


# ================================
# Utilities
# ================================

python-dateutil>=2.8.2
# 时间解析与时区处理
# - 解析网站发布时间
# - 处理非标准日期格式

chardet>=5.2.0
# 字符编码识别
# - 应对老站点 / GBK 编码页面
# - requests 编码兜底方案

emoji>=2.8.0
# Emoji 处理（可选）
# - 用于消息推送美化
# - 不影响核心功能


# ================================
# AI / LLM Support
# ================================

anthropic>=0.3.0
# Anthropic SDK
# - 用于调用 MiniMax 等兼容模型

